+++ app/controllers/search_page_controller.rb
@@ -1,4 +1,0 @@
-class SearchPageController < CmsController
-  def index
-  end
-end